% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{humanicheck}
\alias{humanicheck}
\title{Generate a Data Quality Report}
\usage{
humanicheck(data)
}
\arguments{
\item{data}{The input data frame, tibble, or sf object to be checked.}
}
\value{
The function does not return a value. It prints a formatted report
directly to the console.
}
\description{
This function performs a series of checks on a data frame or tibble and
prints a summary report to the console. The checks include verifying if the
input is a tibble, checking the object's memory size, identifying
duplicated rows, locating NA values, and for \code{sf} objects, ensuring
geometry type consistency.
}
\examples{
\dontrun{
humanicheck(penguins)
}
}
